refactor(#170): вынести форматирование дат и auth-ошибки в утилиты
DateFormatUtils.kt: formatEventDate / formatEventDateFull / formatEventTime
AuthErrorUtils.kt: Exception.toAuthErrorMessage(fallback)

Удалены дублирующие приватные функции из EventDetailScreen и ScannerScreen,
инлайн-форматирование заменено в OrderConfirmScreen,
обработка ошибок унифицирована в LoginScreen, RegisterScreen,
SmsCodeScreen, NameInputScreen.
