+++ composeApp/src/commonMain/kotlin/com/karrad/ticketsclient/ui/screen/auth/LoginScreen.kt
@@ -42,6 +42,7 @@ import cafe.adriel.voyager.navigator.currentOrThrow
 import com.karrad.ticketsclient.crash.CrashReporter
 import com.karrad.ticketsclient.data.api.ApiException
 import com.karrad.ticketsclient.di.AppContainer
+import com.karrad.ticketsclient.ui.util.toAuthErrorMessage
 import com.karrad.ticketsclient.ui.navigation.MainScreen
 import com.karrad.ticketsclient.ui.navigation.RegisterScreen
 import com.karrad.ticketsclient.ui.navigation.SmsCodeScreen
@@ -128,10 +129,10 @@ fun LoginScreen() {
                             AppContainer.authService.sendCode(normalized)
                             navigator.push(SmsCodeScreen(isRegistration = false, phone = normalized))
                         } catch (e: ApiException) {
-                            error = e.message ?: "Не удалось отправить код. Проверьте номер."
+                            error = e.toAuthErrorMessage("Не удалось отправить код. Проверьте номер.")
                         } catch (e: Exception) {
                             CrashReporter.log(e)
-                            error = "Не удалось отправить код. Проверьте номер."
+                            error = e.toAuthErrorMessage("Не удалось отправить код. Проверьте номер.")
                         } finally {
                             loading = false
                         }
