+++ composeApp/src/commonMain/kotlin/com/karrad/ticketsclient/ui/util/AuthErrorUtils.kt
@@ -1,0 +1,4 @@
+package com.karrad.ticketsclient.ui.util
+
+/** Возвращает читаемое сообщение об ошибке из исключения или запасной текст. */
+fun Exception.toAuthErrorMessage(fallback: String): String = message?.takeIf { it.isNotBlank() } ?: fallback
